feat: move data and style to shared
