+++ shared/style.ts
